refactor(#41): Atualiza DTOs para refletir mudanças no modelo dimensional
- Remove campos de projeto da DimAtividadeDTO
- Simplifica DimDevDTO mantendo apenas nome e custo_hora
- Adiciona campos Jira ID nos DTOs de Status e Tipo
- Atualiza FatoApontamentoHorasDTO para usar objetos de dimensão
- Converte ProjectAtividadeCountDTO para usar BigDecimal
- Remove devEmail do DevHoursMetricsDTO
- Atualiza anotações @Schema para RequiredMode.REQUIRED

+++ api/src/main/java/com/neohorizon/api/dto/metrica/ProjectAtividadeCountDTO.java
@@ -1,5 +1,7 @@
 package com.neohorizon.api.dto.metrica;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,5 +11,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ProjectAtividadeCountDTO {
     private String projectName;
-    private long totalAtividades;
+    private BigDecimal totalAtividades;
 }
